requirements.txt
numpy>=1.21.0
pandas>=1.3.0
cvxpy>=1.2.0
pytest>=7.0.0
pyyaml>=6.0.0
pydantic>=2.5.0

# Development dependencies
black>=24.1.0
isort>=5.12.0
mypy>=1.8.0
pytest-cov>=4.1.0  # For test coverage reports

click~=8.1.8
matplotlib~=3.10.1